Add files via upload
Checked L1 and Q1 against recommendations in the datasheet.  Decided to keep as is.  The MOSFET is fine.  L1 should be too and it is rated for 400V. 
Added feed forward capacitor C9, 47pF in parallel with R2.    The datasheet suggests values between 47pF and 220pF, using the lowest that insures stability.  The design is stable in PSPICE without the cap.  In practice it may be optional. 

Updated the PCB with the parts changes and fixed a layout error from Rev 0.